v3.7.0: Split PlannerService and Loc.cs into domain partial files
- PlannerService 1284\u21921388 \u2192 631 lines (Core) + 4 partial files:
  - PlannerService.Finance.cs (547 lines: categories, entries, budgets, recurring, accounts, transfers, goals, analytics, forecast)
  - PlannerService.Debt.cs (53 lines: debts + payments)
  - PlannerService.Inbox.cs (133 lines: inbox + Trello sync)
  - PlannerService.Income.cs (91 lines: income sources + payments)
- Loc.cs 1850 \u2192 128 lines (core) + 4 per-language files:
  - Loc.Ru.cs / Loc.En.cs / Loc.Es.cs / Loc.Fr.cs (as internal static Build() builders)
- Class marked partial, no public API changes, all 20 tests still green
