+++ DailyPlanner/Services/Loc.En.cs
@@ -1,0 +1,432 @@
+namespace DailyPlanner.Services;
+
+internal static class LocEnStrings
+{
+    public static Dictionary<string, string> Build() => new()
+    {
+            ["Today"] = "TODAY",
+            ["TodayBtn"] = "Today",
+            ["Statistics"] = "Statistics",
+            ["Export"] = "Export",
+            ["Settings"] = "Settings",
+            ["Pomodoro"] = "Pomodoro",
+
+            ["TrayOpen"] = "Open",
+            ["TrayToday"] = "Today",
+            ["TrayExit"] = "Exit",
+            ["TrayMinimized"] = "App minimized to tray",
+
+            ["SearchTooltip"] = "Search (Ctrl+F)",
+            ["CopyWeekTooltip"] = "Copy last week's structure",
+
+            ["NoTasksToday"] = "No tasks for today",
+            ["NoTasks"] = "No tasks",
+            ["Completed"] = "completed",
+
+            ["TotalTasks"] = "Total tasks",
+            ["CompletedLabel"] = "Completed",
+            ["NotCompleted"] = "Not completed",
+            ["GoalsReached"] = "Goals reached",
+            ["TotalTasksTip"] = "Number of tasks with text for the week",
+            ["CompletedTip"] = "Tasks marked as completed",
+            ["NotCompletedTip"] = "Tasks not yet completed",
+            ["GoalsTip"] = "Completed goals out of 4",
+
+            ["WeekGoals"] = "WEEK GOALS",
+            ["Notes"] = "NOTES",
+            ["Progress"] = "PROGRESS",
+            ["ProgressLabel"] = "progress",
+            ["BestDay"] = "best day",
+            ["GoalsLabel"] = "goals",
+            ["WeekTasks"] = "WEEKLY TASKS",
+            ["TodayBadge"] = "TODAY",
+            ["Remaining"] = "remaining",
+            ["CarryOverTip"] = "Carry over unfinished to tomorrow",
+            ["AddSubTaskTip"] = "Add subtask",
+            ["DeleteTaskTip"] = "Delete task",
+            ["MoveNextDayTip"] = "Move to next day",
+            ["HabitTracker"] = "HABIT TRACKER",
+            ["StateTracker"] = "WELL-BEING",
+            ["StateNoData"] = "Track your state each day",
+            ["StateBestDay"] = "Best day: {0}",
+            ["StateWorstDay"] = "Tough day: {0}",
+            ["StateAvg"] = "Average",
+            ["AddNoteTip"] = "Add note",
+            ["AddGoalTip"] = "Add goal",
+            ["DeleteTip"] = "Delete",
+            ["AddHabitTip"] = "Add habit",
+            ["HabitEmpty"] = "Press + to add a habit",
+            ["HabitLabel"] = "Habit",
+            ["ProgressCol"] = "Progress",
+
+            ["Monday"] = "Monday", ["Tuesday"] = "Tuesday", ["Wednesday"] = "Wednesday",
+            ["Thursday"] = "Thursday", ["Friday"] = "Friday", ["Saturday"] = "Saturday", ["Sunday"] = "Sunday",
+            ["Mon"] = "Mo", ["Tue"] = "Tu", ["Wed"] = "We", ["Thu"] = "Th",
+            ["Fri"] = "Fr", ["Sat"] = "Sa", ["Sun"] = "Su",
+
+            ["Sleep"] = "Sleep", ["Energy"] = "Energy", ["Mood"] = "Mood",
+
+            ["Month1"] = "January", ["Month2"] = "February", ["Month3"] = "March",
+            ["Month4"] = "April", ["Month5"] = "May", ["Month6"] = "June",
+            ["Month7"] = "July", ["Month8"] = "August", ["Month9"] = "September",
+            ["Month10"] = "October", ["Month11"] = "November", ["Month12"] = "December",
+
+            ["PomWork"] = "Work", ["PomBreak"] = "Break", ["PomFocus"] = "Focus",
+            ["PomPomodoro"] = "Pomodoro", ["PomFocusMode"] = "Focus mode",
+            ["PomReset"] = "Reset", ["PomSkip"] = "Skip",
+            ["PomToggleFocus"] = "Toggle Pomodoro/Focus",
+            ["PomSessions"] = "Sessions: ",
+            ["PomPause"] = "\u23F8 Pause", ["PomStart"] = "\u25B6 Start",
+            ["PomBreakTime"] = "Break time! Rest up \U0001f3d6\ufe0f",
+            ["PomWorkTime"] = "Time to work! \U0001f4aa",
+            ["PomFocusTimer"] = "Focus Timer",
+            ["PomFocusAlert"] = "Already {0} min! Take a break \u2615",
+
+            ["PriorityHigh"] = "High priority (click to change)",
+            ["PriorityMedium"] = "Medium priority (click to change)",
+            ["PriorityLow"] = "Low priority (click to change)",
+            ["PriorityNone"] = "No priority",
+
+            ["CatWork"] = "Work (click to change)", ["CatStudy"] = "Study (click to change)",
+            ["CatPersonal"] = "Personal (click to change)", ["CatHealthTip"] = "Health (click to change)",
+            ["CatOther"] = "Other (click to change)",
+
+            ["SettingsTitle"] = "Settings",
+            ["ExportExcel"] = "Export to Excel",
+            ["ExportExcelDesc"] = "Export the current week to .xlsx file",
+            ["Palette"] = "Palette",
+            ["PaletteCurrent"] = "Current: ",
+            ["AutoStart"] = "Auto start",
+            ["AutoStartDesc"] = "Launch the app on Windows login",
+            ["Updates"] = "Updates",
+            ["CheckUpdates"] = "Check",
+            ["UpdateBtn"] = "Update",
+            ["Backup"] = "Backup",
+            ["BackupDesc"] = "Create a copy or restore the database",
+            ["CreateBackup"] = "Create backup",
+            ["Restore"] = "Restore",
+            ["Reminders"] = "Reminders",
+            ["RemindersDesc"] = "Notifications at specified time",
+            ["Add"] = "Add",
+            ["TimeTip"] = "Time (HH:MM)",
+            ["TitleTip"] = "Title",
+            ["MessageTip"] = "Message",
+            ["RecurringTasks"] = "Recurring tasks",
+            ["RecurringDesc"] = "Templates are automatically added every week",
+            ["AboutApp"] = "About",
+            ["Version"] = "Version",
+            ["Platform"] = "Platform",
+            ["Database"] = "Database",
+            ["DbLocal"] = "SQLite (local)",
+            ["Storage"] = "Storage",
+            ["Shortcuts"] = "Keyboard shortcuts",
+            ["SwitchMonth"] = "Switch month",
+            ["GoToToday"] = "Go to today",
+            ["SearchTasks"] = "Search tasks",
+            ["PomodoroTimer"] = "Pomodoro timer",
+            ["ExportExcelShort"] = "Export to Excel",
+            ["OpenSettings"] = "Open settings",
+            ["Description"] = "Description",
+            ["AppDescription"] = "Daily & Financial Planner — a weekly planner with finance module, habit tracker, productivity analytics, and state monitoring. All data is stored locally.",
+            ["Language"] = "Language",
+            ["LanguageDesc"] = "Application interface language",
+
+            ["Productivity"] = "Productivity",
+            ["BestDayLabel"] = "Best day",
+            ["Goals"] = "Goals",
+            ["AvgSleep"] = "Avg. sleep",
+            ["AvgEnergy"] = "Avg. energy",
+            ["AvgMood"] = "Avg. mood",
+            ["WeekComparison"] = "Week comparison",
+            ["PrevWeek"] = "Previous week",
+            ["Trend"] = "Trend",
+            ["CurrWeek"] = "Current week",
+            ["WeeklyProductivity"] = "Weekly productivity",
+            ["HabitHeatmap"] = "Habits — Heatmap (30 days)",
+            ["HabitStreaks"] = "Habit streaks",
+            ["TasksUnit"] = "tasks",
+            ["NoChange"] = "No change",
+            ["DaysUnit"] = "d.",
+
+            ["SummaryTasks"] = "Tasks: {0}/{1}",
+            ["SummaryGoals"] = "Goals: {0}/{1}",
+            ["SummaryBestDay"] = "Best day: {0}",
+            ["SummaryExcellent"] = "\u2b50 Excellent week!",
+            ["SummaryGood"] = "\u2705 Good week!",
+            ["SummaryCanBetter"] = "\U0001f4aa Can do better!",
+
+            ["ExportTitle"] = "Export",
+            ["ExportSuccess"] = "File saved successfully",
+            ["ExportError"] = "Export error",
+            ["RestoreTitle"] = "Restore",
+            ["RestoreSuccess"] = "Database restored successfully",
+            ["RestoreError"] = "Error restoring database",
+            ["RestoreInvalidDb"] = "File is not a valid SQLite database",
+            ["CopyTitle"] = "Copy",
+            ["CopySuccess"] = "Last week's structure copied",
+            ["Reminder"] = "Reminder",
+            ["MoveTitle"] = "Move",
+            ["MoveNoSlots"] = "No empty slots in this day",
+
+            ["UpdateUnavailable"] = "Updates unavailable (dev mode)",
+            ["UpdateChecking"] = "Checking for updates...",
+            ["UpdateLatest"] = "You are using the latest version",
+            ["UpdateAvailable"] = "Update available: v{0}",
+            ["UpdateDownloading"] = "Downloading update...",
+            ["UpdateProgress"] = "Downloading: {0}%",
+
+            ["ExpWeek"] = "Week {0}",
+            ["ExpPlannerWeek"] = "Planner — Week {0} – {1}",
+            ["ExpWeekGoals"] = "Week goals",
+            ["ExpDay"] = "Day",
+            ["ExpTask"] = "Task {0}",
+            ["ExpState"] = "State",
+            ["ExpHabits"] = "Habits",
+            ["ExpNotes"] = "Notes",
+
+            ["Loading"] = "Loading data...",
+            ["DbError"] = "Database initialization error:\n{0}",
+            ["PlannerFileName"] = "Planner_{0}",
+            ["ExcelFilter"] = "Excel files (*.xlsx)|*.xlsx",
+            ["Goal"] = "Goal",
+            ["LightTheme"] = "Light",
+
+            ["Quote1"] = "Small steps every day lead to big results.",
+            ["Quote2"] = "Discipline is the bridge between goals and accomplishment.",
+            ["Quote3"] = "Don't wait for the perfect moment. Take the moment and make it perfect.",
+            ["Quote4"] = "Success is the sum of small efforts repeated day in and day out.",
+            ["Quote5"] = "Planning is bringing the future into the present.",
+            ["Quote6"] = "Every day is a new opportunity to become better.",
+            ["Quote7"] = "Do today what others won't. Tomorrow you'll live as others can't.",
+            ["Quote8"] = "Focus is not about saying 'yes', it's about saying 'no'.",
+            ["Quote9"] = "Habits define the future. Choose them wisely.",
+            ["Quote10"] = "Progress, not perfection.",
+            ["Quote11"] = "The best time to start is right now.",
+            ["Quote12"] = "Your only competitor is who you were yesterday.",
+            ["Quote13"] = "Energy follows attention. Focus on what matters.",
+            ["Quote14"] = "Without a plan you are planning to fail.",
+            ["Quote15"] = "Great things start small.",
+            ["Quote16"] = "Focus on the process and results will follow.",
+            ["Quote17"] = "Every completed task is a small victory.",
+            ["Quote18"] = "Don't put off until tomorrow what will make you better today.",
+            ["Quote19"] = "Productivity is not busyness, it's results.",
+            ["Quote20"] = "One hour of planning saves ten hours of work.",
+
+            // Meetings
+            ["Meetings"] = "Meetings",
+            ["MeetingsDesc"] = "Meeting scheduler with reminders",
+            ["NewMeeting"] = "New meeting",
+            ["MeetingTitle"] = "Title",
+            ["MeetingDesc"] = "Description",
+            ["MeetingAttendees"] = "Attendees",
+            ["MeetingDate"] = "Date",
+            ["MeetingTime"] = "Time",
+            ["MeetingDuration"] = "Duration",
+            ["MeetingMin"] = "min",
+            ["MeetingNotifyDay"] = "1 day before",
+            ["MeetingNotify2h"] = "2 hours before",
+            ["MeetingNotify30m"] = "30 min before",
+            ["MeetingTomorrow"] = "Meeting tomorrow",
+            ["MeetingSoon"] = "Meeting soon",
+            ["MeetingIn2Hours"] = "in 2 hours",
+            ["MeetingIn30Min"] = "in 30 minutes",
+            ["MeetingNoMeetings"] = "No scheduled meetings",
+            ["MeetingUpcoming"] = "Upcoming",
+            ["MeetingPast"] = "Past",
+
+            ["Undo"] = "Undo",
+            ["TaskDeleted"] = "Task \"{0}\" deleted",
+            ["CtxDuplicate"] = "Duplicate",
+            ["CtxPriority"] = "Priority",
+            ["CtxCategory"] = "Category",
+            ["CtxDeadline"] = "Deadline",
+            ["CtxDeadlineToday"] = "Today",
+            ["CtxDeadlineTomorrow"] = "Tomorrow",
+            ["CtxDeadlineNextWeek"] = "Next week",
+            ["CtxDeadlineClear"] = "Clear deadline",
+            ["CtxMoveNext"] = "Move to next day",
+            ["CtxDelete"] = "Delete",
+            ["CtxAddSub"] = "Add subtask",
+            ["CtxNone"] = "None",
+            ["Overdue"] = "Overdue",
+            ["DaysShort"] = "d",
+            ["DueToday"] = "Today",
+            ["DueTomorrow"] = "Tomorrow",
+            ["MyDayTitle"] = "My Day",
+            ["MyDayGreeting"] = "Good morning!",
+            ["GreetingMorning"] = "Good morning!",
+            ["GreetingAfternoon"] = "Good afternoon!",
+            ["GreetingEvening"] = "Good evening!",
+            ["GreetingNight"] = "Good night!",
+            ["MyDayYesterday"] = "Not finished yesterday:",
+            ["MyDayToday"] = "Today's tasks:",
+            ["MyDayOverdue"] = "Overdue tasks:",
+            ["MyDayNoTasks"] = "No tasks for today. Time to plan!",
+            ["MyDayStart"] = "Start the day",
+            ["MyDayDontShow"] = "Don't show today",
+            ["MyDayEmpty"] = "New day — new opportunities!",
+            ["MyDaySummaryTasks"] = "Tasks: {0} (done {1})",
+            ["MyDaySummaryIncomplete"] = "Unfinished yesterday: {0}",
+            ["MyDaySummaryOverdue"] = "Overdue: {0}",
+
+            // Finance
+            ["Finance"] = "Finance",
+            ["FinanceTooltip"] = "Finance (Ctrl+M)",
+            ["Income"] = "Income",
+            ["Expenses"] = "Expenses",
+            ["BalanceLabel"] = "Balance",
+            ["Debts"] = "Debts",
+            ["RecurringPayments"] = "Recurring",
+            ["AddEntry"] = "Add entry",
+            ["AddDebt"] = "Add debt",
+            ["AddPayment"] = "Add payment",
+            ["AddBudget"] = "Add budget",
+            ["AddRecurring"] = "Add payment",
+            ["Amount"] = "Amount",
+            ["Person"] = "Person",
+            ["DebtLent"] = "Owed to me",
+            ["DebtBorrowed"] = "I owe",
+            ["DebtSettled"] = "Settled",
+            ["FinRemaining"] = "Remaining",
+            ["Budget"] = "Budget",
+            ["Spent"] = "Spent",
+            ["OverBudget"] = "Over budget",
+            ["FreqMonthly"] = "Monthly",
+            ["FreqWeekly"] = "Weekly",
+            ["FreqBiweekly"] = "Biweekly",
+            ["FreqQuarterly"] = "Quarterly",
+            ["FreqYearly"] = "Yearly",
+            ["DayOfMonthSuffix"] = "th",
+            ["NoEntries"] = "No entries",
+            ["NoDebts"] = "No debts",
+            ["FinOverdue"] = "Overdue",
+            ["Active"] = "Active",
+            ["Auto"] = "Auto",
+            ["Remind"] = "Remind",
+            ["MonthlyObligatory"] = "Obligatory/mo",
+            ["FreeAmount"] = "Free",
+            ["Paid"] = "Paid",
+            ["Pending"] = "Pending",
+            ["FinanceDescription"] = "Description",
+            ["Savings"] = "Savings",
+            ["SavingsRate"] = "Savings rate",
+            ["NetWorth"] = "Net Worth",
+            ["ChangeIcon"] = "Change icon",
+            ["IncomeSources"] = "Income Sources",
+            ["ExpectedIncome"] = "Expected Income",
+            ["ReceivedIncome"] = "Received",
+            ["ProjectName"] = "Project name",
+            ["ClientName"] = "Client",
+            ["TotalMonthly"] = "Total/month",
+            ["PaymentDay"] = "Payment day",
+            ["AddSource"] = "Add source",
+            ["AddPaymentDate"] = "Add payment date",
+            ["Scheduled"] = "Scheduled",
+            ["Paid"] = "Paid",
+            ["Pending"] = "Pending",
+            ["PaymentSchedule"] = "Payment schedule",
+            ["Analytics"] = "Analytics",
+            ["FinCategories"] = "By category",
+            ["MonthlyTrend"] = "Monthly trend",
+            ["ExportFinance"] = "Export finance",
+            ["AnnualTotal"] = "Per year",
+            ["NextPayment"] = "Next payment",
+            ["Date"] = "Date",
+            // Finance categories
+            ["CatSalary"] = "Salary",
+            ["CatFreelance"] = "Freelance",
+            ["CatGifts"] = "Gifts",
+            ["CatInvestments"] = "Investments",
+            ["CatOtherIncome"] = "Other Income",
+            ["CatFood"] = "Food",
+            ["CatTransport"] = "Transport",
+            ["CatHousing"] = "Housing",
+            ["CatEntertainment"] = "Entertainment",
+            ["CatHealth"] = "Health",
+            ["CatClothing"] = "Clothing",
+            ["CatSubscriptions"] = "Subscriptions",
+            ["CatEducation"] = "Education",
+            ["CatOtherExpense"] = "Other Expense",
+            // Financial Goals
+            ["FinGoals"] = "Financial Goals",
+            ["AddFinGoal"] = "Add goal",
+            ["TargetAmount"] = "Target amount",
+            ["SavedAmount"] = "Saved",
+            ["TargetDate"] = "Target date",
+            ["GoalReached"] = "Goal reached!",
+            ["NoFinGoals"] = "No financial goals",
+            ["GoalProgress"] = "{0}% of goal",
+            // Accounts
+            ["Accounts"] = "Accounts",
+            ["AddAccount"] = "Add account",
+            ["Transfer"] = "Transfer",
+            ["AddTransfer"] = "Add transfer",
+            ["TransferFrom"] = "From account",
+            ["TransferTo"] = "To account",
+            ["InitialBalance"] = "Initial balance",
+            ["AccountBalance"] = "Account balance",
+            ["NoAccounts"] = "No accounts",
+            // Filter / Search
+            ["FilterEntries"] = "Filter entries",
+            ["SearchPlaceholder"] = "Search...",
+            ["AllCategories"] = "All categories",
+            ["DateFrom"] = "From date",
+            ["DateTo"] = "To date",
+            ["ClearFilter"] = "Clear",
+            // Forecast
+            ["Forecast"] = "Forecast",
+            ["ForecastBalance"] = "Balance forecast",
+            ["ProjectedIncome"] = "Projected income",
+            ["ProjectedExpenses"] = "Projected expenses",
+            ["ForecastDays"] = "for 30 days",
+            // Cashflow
+            ["Cashflow"] = "Cash flow",
+            ["CashflowCalendar"] = "Payment calendar",
+            // Import
+            ["ImportFile"] = "Import from file",
+            ["SelectFile"] = "Select file",
+            ["ImportSuccess"] = "{0} entries imported",
+            ["ImportError"] = "Import error",
+            ["ImportCSV"] = "CSV files|*.csv",
+            ["ImportExcel"] = "Excel files|*.xlsx",
+            // Split
+            ["SplitTransaction"] = "Split transaction",
+            ["AddSplit"] = "Add split",
+            ["SplitTotal"] = "Split total",
+            ["SplitParent"] = "Parent entry",
+            // Trends
+            ["TrendChart"] = "Trends",
+            ["IncomeVsExpenses"] = "Income vs Expenses",
+            ["DueDate"] = "Due date",
+            ["SetDueDate"] = "Set due date",
+            ["ClearDate"] = "Clear date",
+            ["PickDate"] = "Pick date",
+
+            // Inbox / Trello
+            ["InboxTitle"] = "Inbox",
+            ["InboxEmpty"] = "No inbox tasks",
+            ["InboxEmptyHint"] = "Add manually or sync with Trello",
+            ["InboxMovedToDate"] = "Task moved to {0}",
+            ["AddTask"] = "Add",
+            ["SyncTrello"] = "Sync Trello",
+            ["MoveToDay"] = "Move to day",
+            ["Delete"] = "Delete",
+            ["NewTask"] = "New task",
+            ["TrelloIntegration"] = "Trello Integration",
+            ["TrelloDesc"] = "Pulls cards from the specified list into the planner inbox.",
+            ["TrelloApiKey"] = "API Key",
+            ["TrelloToken"] = "Token",
+            ["TrelloListName"] = "List name (e.g. \"Doing\")",
+            ["TrelloHelp"] = "Get key at https://trello.com/app-key — same page has a \"Token\" button.",
+            ["TestConnection"] = "Test connection",
+            ["TrelloTesting"] = "Testing...",
+            ["TrelloConnectionOk"] = "Connection successful",
+            ["TrelloConnectionFail"] = "Connection failed",
+            ["TrelloSaved"] = "Settings saved",
+            ["TrelloSyncing"] = "Syncing...",
+            ["TrelloSyncDone"] = "Added {0} cards",
+            ["TrelloSyncError"] = "Sync error",
+            ["Enabled"] = "Enabled",
+            ["Save"] = "Save",
+    };
+}
