+++ DailyPlanner/Services/Loc.Es.cs
@@ -1,0 +1,432 @@
+namespace DailyPlanner.Services;
+
+internal static class LocEsStrings
+{
+    public static Dictionary<string, string> Build() => new()
+    {
+            ["Today"] = "HOY",
+            ["TodayBtn"] = "Hoy",
+            ["Statistics"] = "Estadísticas",
+            ["Export"] = "Exportar",
+            ["Settings"] = "Configuración",
+            ["Pomodoro"] = "Pomodoro",
+
+            ["TrayOpen"] = "Abrir",
+            ["TrayToday"] = "Hoy",
+            ["TrayExit"] = "Salir",
+            ["TrayMinimized"] = "Aplicación minimizada a la bandeja",
+
+            ["SearchTooltip"] = "Buscar (Ctrl+F)",
+            ["CopyWeekTooltip"] = "Copiar estructura de la semana pasada",
+
+            ["NoTasksToday"] = "Sin tareas para hoy",
+            ["NoTasks"] = "Sin tareas",
+            ["Completed"] = "completadas",
+
+            ["TotalTasks"] = "Total de tareas",
+            ["CompletedLabel"] = "Completadas",
+            ["NotCompleted"] = "No completadas",
+            ["GoalsReached"] = "Metas alcanzadas",
+            ["TotalTasksTip"] = "Número de tareas con texto en la semana",
+            ["CompletedTip"] = "Tareas marcadas como completadas",
+            ["NotCompletedTip"] = "Tareas aún no completadas",
+            ["GoalsTip"] = "Metas completadas de 4",
+
+            ["WeekGoals"] = "METAS DE LA SEMANA",
+            ["Notes"] = "NOTAS",
+            ["Progress"] = "PROGRESO",
+            ["ProgressLabel"] = "progreso",
+            ["BestDay"] = "mejor día",
+            ["GoalsLabel"] = "metas",
+            ["WeekTasks"] = "TAREAS DE LA SEMANA",
+            ["TodayBadge"] = "HOY",
+            ["Remaining"] = "restantes",
+            ["CarryOverTip"] = "Pasar pendientes a mañana",
+            ["AddSubTaskTip"] = "Agregar subtarea",
+            ["DeleteTaskTip"] = "Eliminar tarea",
+            ["MoveNextDayTip"] = "Mover al día siguiente",
+            ["HabitTracker"] = "SEGUIMIENTO DE HÁBITOS",
+            ["StateTracker"] = "BIENESTAR",
+            ["StateNoData"] = "Registra tu estado cada día",
+            ["StateBestDay"] = "Mejor día: {0}",
+            ["StateWorstDay"] = "Día difícil: {0}",
+            ["StateAvg"] = "Promedio",
+            ["AddNoteTip"] = "Añadir nota",
+            ["AddGoalTip"] = "Añadir objetivo",
+            ["DeleteTip"] = "Eliminar",
+            ["AddHabitTip"] = "Añadir hábito",
+            ["HabitEmpty"] = "Pulse + para añadir un hábito",
+            ["HabitLabel"] = "Hábito",
+            ["ProgressCol"] = "Progreso",
+
+            ["Monday"] = "Lunes", ["Tuesday"] = "Martes", ["Wednesday"] = "Miércoles",
+            ["Thursday"] = "Jueves", ["Friday"] = "Viernes", ["Saturday"] = "Sábado", ["Sunday"] = "Domingo",
+            ["Mon"] = "Lu", ["Tue"] = "Ma", ["Wed"] = "Mi", ["Thu"] = "Ju",
+            ["Fri"] = "Vi", ["Sat"] = "Sá", ["Sun"] = "Do",
+
+            ["Sleep"] = "Sueño", ["Energy"] = "Energía", ["Mood"] = "Ánimo",
+
+            ["Month1"] = "Enero", ["Month2"] = "Febrero", ["Month3"] = "Marzo",
+            ["Month4"] = "Abril", ["Month5"] = "Mayo", ["Month6"] = "Junio",
+            ["Month7"] = "Julio", ["Month8"] = "Agosto", ["Month9"] = "Septiembre",
+            ["Month10"] = "Octubre", ["Month11"] = "Noviembre", ["Month12"] = "Diciembre",
+
+            ["PomWork"] = "Trabajo", ["PomBreak"] = "Descanso", ["PomFocus"] = "Enfoque",
+            ["PomPomodoro"] = "Pomodoro", ["PomFocusMode"] = "Modo enfoque",
+            ["PomReset"] = "Reiniciar", ["PomSkip"] = "Saltar",
+            ["PomToggleFocus"] = "Alternar Pomodoro/Enfoque",
+            ["PomSessions"] = "Sesiones: ",
+            ["PomPause"] = "\u23F8 Pausa", ["PomStart"] = "\u25B6 Iniciar",
+            ["PomBreakTime"] = "¡Hora del descanso! Relájate \U0001f3d6\ufe0f",
+            ["PomWorkTime"] = "¡Hora de trabajar! \U0001f4aa",
+            ["PomFocusTimer"] = "Temporizador de enfoque",
+            ["PomFocusAlert"] = "¡Ya {0} min! Toma un descanso \u2615",
+
+            ["PriorityHigh"] = "Prioridad alta (clic para cambiar)",
+            ["PriorityMedium"] = "Prioridad media (clic para cambiar)",
+            ["PriorityLow"] = "Prioridad baja (clic para cambiar)",
+            ["PriorityNone"] = "Sin prioridad",
+
+            ["CatWork"] = "Trabajo (clic para cambiar)", ["CatStudy"] = "Estudio (clic para cambiar)",
+            ["CatPersonal"] = "Personal (clic para cambiar)", ["CatHealthTip"] = "Salud (clic para cambiar)",
+            ["CatOther"] = "Otro (clic para cambiar)",
+
+            ["SettingsTitle"] = "Configuración",
+            ["ExportExcel"] = "Exportar a Excel",
+            ["ExportExcelDesc"] = "Exportar la semana actual a archivo .xlsx",
+            ["Palette"] = "Paleta",
+            ["PaletteCurrent"] = "Actual: ",
+            ["AutoStart"] = "Inicio automático",
+            ["AutoStartDesc"] = "Iniciar la aplicación al entrar en Windows",
+            ["Updates"] = "Actualizaciones",
+            ["CheckUpdates"] = "Comprobar",
+            ["UpdateBtn"] = "Actualizar",
+            ["Backup"] = "Copia de seguridad",
+            ["BackupDesc"] = "Crear una copia o restaurar la base de datos",
+            ["CreateBackup"] = "Crear copia",
+            ["Restore"] = "Restaurar",
+            ["Reminders"] = "Recordatorios",
+            ["RemindersDesc"] = "Notificaciones a la hora indicada",
+            ["Add"] = "Añadir",
+            ["TimeTip"] = "Hora (HH:MM)",
+            ["TitleTip"] = "Título",
+            ["MessageTip"] = "Mensaje",
+            ["RecurringTasks"] = "Tareas recurrentes",
+            ["RecurringDesc"] = "Las plantillas se añaden automáticamente cada semana",
+            ["AboutApp"] = "Acerca de",
+            ["Version"] = "Versión",
+            ["Platform"] = "Plataforma",
+            ["Database"] = "Base de datos",
+            ["DbLocal"] = "SQLite (local)",
+            ["Storage"] = "Almacenamiento",
+            ["Shortcuts"] = "Atajos de teclado",
+            ["SwitchMonth"] = "Cambiar mes",
+            ["GoToToday"] = "Ir a hoy",
+            ["SearchTasks"] = "Buscar tareas",
+            ["PomodoroTimer"] = "Temporizador Pomodoro",
+            ["ExportExcelShort"] = "Exportar a Excel",
+            ["OpenSettings"] = "Abrir configuración",
+            ["Description"] = "Descripción",
+            ["AppDescription"] = "Daily & Financial Planner — planificador semanal con módulo financiero, seguimiento de hábitos, análisis de productividad y monitoreo de estado. Todos los datos se almacenan localmente.",
+            ["Language"] = "Idioma",
+            ["LanguageDesc"] = "Idioma de la interfaz de la aplicación",
+
+            ["Productivity"] = "Productividad",
+            ["BestDayLabel"] = "Mejor día",
+            ["Goals"] = "Metas",
+            ["AvgSleep"] = "Prom. sueño",
+            ["AvgEnergy"] = "Prom. energía",
+            ["AvgMood"] = "Prom. ánimo",
+            ["WeekComparison"] = "Comparación semanal",
+            ["PrevWeek"] = "Semana anterior",
+            ["Trend"] = "Tendencia",
+            ["CurrWeek"] = "Semana actual",
+            ["WeeklyProductivity"] = "Productividad semanal",
+            ["HabitHeatmap"] = "Hábitos — Mapa de calor (30 días)",
+            ["HabitStreaks"] = "Rachas de hábitos",
+            ["TasksUnit"] = "tareas",
+            ["NoChange"] = "Sin cambios",
+            ["DaysUnit"] = "d.",
+
+            ["SummaryTasks"] = "Tareas: {0}/{1}",
+            ["SummaryGoals"] = "Metas: {0}/{1}",
+            ["SummaryBestDay"] = "Mejor día: {0}",
+            ["SummaryExcellent"] = "\u2b50 ¡Semana excelente!",
+            ["SummaryGood"] = "\u2705 ¡Buena semana!",
+            ["SummaryCanBetter"] = "\U0001f4aa ¡Se puede mejorar!",
+
+            ["ExportTitle"] = "Exportar",
+            ["ExportSuccess"] = "Archivo guardado correctamente",
+            ["ExportError"] = "Error al exportar",
+            ["RestoreTitle"] = "Restaurar",
+            ["RestoreSuccess"] = "Base de datos restaurada correctamente",
+            ["RestoreError"] = "Error al restaurar la base de datos",
+            ["RestoreInvalidDb"] = "El archivo no es una base de datos SQLite válida",
+            ["CopyTitle"] = "Copiar",
+            ["CopySuccess"] = "Estructura de la semana pasada copiada",
+            ["Reminder"] = "Recordatorio",
+            ["MoveTitle"] = "Mover",
+            ["MoveNoSlots"] = "No hay espacios libres en este día",
+
+            ["UpdateUnavailable"] = "Actualizaciones no disponibles (modo dev)",
+            ["UpdateChecking"] = "Comprobando actualizaciones...",
+            ["UpdateLatest"] = "Estás usando la última versión",
+            ["UpdateAvailable"] = "Actualización disponible: v{0}",
+            ["UpdateDownloading"] = "Descargando actualización...",
+            ["UpdateProgress"] = "Descargando: {0}%",
+
+            ["ExpWeek"] = "Semana {0}",
+            ["ExpPlannerWeek"] = "Planificador — Semana {0} – {1}",
+            ["ExpWeekGoals"] = "Metas de la semana",
+            ["ExpDay"] = "Día",
+            ["ExpTask"] = "Tarea {0}",
+            ["ExpState"] = "Estado",
+            ["ExpHabits"] = "Hábitos",
+            ["ExpNotes"] = "Notas",
+
+            ["Loading"] = "Cargando datos...",
+            ["DbError"] = "Error al inicializar la base de datos:\n{0}",
+            ["PlannerFileName"] = "Planificador_{0}",
+            ["ExcelFilter"] = "Archivos Excel (*.xlsx)|*.xlsx",
+            ["Goal"] = "Meta",
+            ["LightTheme"] = "Claro",
+
+            ["Quote1"] = "Pequeños pasos cada día llevan a grandes resultados.",
+            ["Quote2"] = "La disciplina es el puente entre las metas y los logros.",
+            ["Quote3"] = "No esperes el momento perfecto. Toma el momento y hazlo perfecto.",
+            ["Quote4"] = "El éxito es la suma de pequeños esfuerzos repetidos día tras día.",
+            ["Quote5"] = "Planificar es traer el futuro al presente.",
+            ["Quote6"] = "Cada día es una nueva oportunidad para ser mejor.",
+            ["Quote7"] = "Haz hoy lo que otros no quieren. Mañana vivirás como otros no pueden.",
+            ["Quote8"] = "El enfoque no es decir 'sí', es decir 'no'.",
+            ["Quote9"] = "Los hábitos definen el futuro. Elígelos sabiamente.",
+            ["Quote10"] = "Progreso, no perfección.",
+            ["Quote11"] = "El mejor momento para empezar es ahora mismo.",
+            ["Quote12"] = "Tu único competidor es quien fuiste ayer.",
+            ["Quote13"] = "La energía sigue a la atención. Enfócate en lo importante.",
+            ["Quote14"] = "Sin un plan estás planeando fracasar.",
+            ["Quote15"] = "Las grandes cosas empiezan en pequeño.",
+            ["Quote16"] = "Concéntrate en el proceso y los resultados vendrán.",
+            ["Quote17"] = "Cada tarea completada es una pequeña victoria.",
+            ["Quote18"] = "No dejes para mañana lo que te hará mejor hoy.",
+            ["Quote19"] = "La productividad no es estar ocupado, es obtener resultados.",
+            ["Quote20"] = "Una hora de planificación ahorra diez horas de trabajo.",
+
+            // Meetings
+            ["Meetings"] = "Reuniones",
+            ["MeetingsDesc"] = "Programador de reuniones con recordatorios",
+            ["NewMeeting"] = "Nueva reunión",
+            ["MeetingTitle"] = "Título",
+            ["MeetingDesc"] = "Descripción",
+            ["MeetingAttendees"] = "Participantes",
+            ["MeetingDate"] = "Fecha",
+            ["MeetingTime"] = "Hora",
+            ["MeetingDuration"] = "Duración",
+            ["MeetingMin"] = "min",
+            ["MeetingNotifyDay"] = "1 día antes",
+            ["MeetingNotify2h"] = "2 horas antes",
+            ["MeetingNotify30m"] = "30 min antes",
+            ["MeetingTomorrow"] = "Reunión mañana",
+            ["MeetingSoon"] = "Reunión pronto",
+            ["MeetingIn2Hours"] = "en 2 horas",
+            ["MeetingIn30Min"] = "en 30 minutos",
+            ["MeetingNoMeetings"] = "No hay reuniones programadas",
+            ["MeetingUpcoming"] = "Próximas",
+            ["MeetingPast"] = "Pasadas",
+
+            ["Undo"] = "Deshacer",
+            ["TaskDeleted"] = "Tarea \"{0}\" eliminada",
+            ["CtxDuplicate"] = "Duplicar",
+            ["CtxPriority"] = "Prioridad",
+            ["CtxCategory"] = "Categoría",
+            ["CtxDeadline"] = "Plazo",
+            ["CtxDeadlineToday"] = "Hoy",
+            ["CtxDeadlineTomorrow"] = "Mañana",
+            ["CtxDeadlineNextWeek"] = "Próxima semana",
+            ["CtxDeadlineClear"] = "Quitar plazo",
+            ["CtxMoveNext"] = "Mover al día siguiente",
+            ["CtxDelete"] = "Eliminar",
+            ["CtxAddSub"] = "Agregar subtarea",
+            ["CtxNone"] = "Ninguno",
+            ["Overdue"] = "Vencido",
+            ["DaysShort"] = "d",
+            ["DueToday"] = "Hoy",
+            ["DueTomorrow"] = "Mañana",
+            ["MyDayTitle"] = "Mi día",
+            ["MyDayGreeting"] = "¡Buenos días!",
+            ["GreetingMorning"] = "¡Buenos días!",
+            ["GreetingAfternoon"] = "¡Buenas tardes!",
+            ["GreetingEvening"] = "¡Buenas tardes!",
+            ["GreetingNight"] = "¡Buenas noches!",
+            ["MyDayYesterday"] = "No terminado ayer:",
+            ["MyDayToday"] = "Tareas de hoy:",
+            ["MyDayOverdue"] = "Tareas vencidas:",
+            ["MyDayNoTasks"] = "Sin tareas para hoy. ¡Hora de planificar!",
+            ["MyDayStart"] = "Empezar el día",
+            ["MyDayDontShow"] = "No mostrar hoy",
+            ["MyDayEmpty"] = "¡Nuevo día — nuevas oportunidades!",
+            ["MyDaySummaryTasks"] = "Tareas: {0} (hechas {1})",
+            ["MyDaySummaryIncomplete"] = "No terminadas ayer: {0}",
+            ["MyDaySummaryOverdue"] = "Vencidas: {0}",
+
+            // Finance
+            ["Finance"] = "Finanzas",
+            ["FinanceTooltip"] = "Finanzas (Ctrl+M)",
+            ["Income"] = "Ingresos",
+            ["Expenses"] = "Gastos",
+            ["BalanceLabel"] = "Balance",
+            ["Debts"] = "Deudas",
+            ["RecurringPayments"] = "Obligatorios",
+            ["AddEntry"] = "Agregar registro",
+            ["AddDebt"] = "Agregar deuda",
+            ["AddPayment"] = "Agregar pago",
+            ["AddBudget"] = "Agregar presupuesto",
+            ["AddRecurring"] = "Agregar pago",
+            ["Amount"] = "Monto",
+            ["Person"] = "Persona",
+            ["DebtLent"] = "Me deben",
+            ["DebtBorrowed"] = "Yo debo",
+            ["DebtSettled"] = "Saldado",
+            ["FinRemaining"] = "Restante",
+            ["Budget"] = "Presupuesto",
+            ["Spent"] = "Gastado",
+            ["OverBudget"] = "Excedido",
+            ["FreqMonthly"] = "Mensual",
+            ["FreqWeekly"] = "Semanal",
+            ["FreqBiweekly"] = "Quincenal",
+            ["FreqQuarterly"] = "Trimestral",
+            ["FreqYearly"] = "Anual",
+            ["DayOfMonthSuffix"] = "o",
+            ["NoEntries"] = "Sin registros",
+            ["NoDebts"] = "Sin deudas",
+            ["FinOverdue"] = "Vencido",
+            ["Active"] = "Activo",
+            ["Auto"] = "Auto",
+            ["Remind"] = "Recordar",
+            ["MonthlyObligatory"] = "Obligatorios/mes",
+            ["FreeAmount"] = "Libre",
+            ["Paid"] = "Pagado",
+            ["Pending"] = "Pendiente",
+            ["FinanceDescription"] = "Descripción",
+            ["Savings"] = "Ahorros",
+            ["SavingsRate"] = "Tasa de ahorro",
+            ["NetWorth"] = "Patrimonio neto",
+            ["ChangeIcon"] = "Cambiar icono",
+            ["IncomeSources"] = "Fuentes de ingreso",
+            ["ExpectedIncome"] = "Ingreso esperado",
+            ["ReceivedIncome"] = "Recibido",
+            ["ProjectName"] = "Nombre del proyecto",
+            ["ClientName"] = "Cliente",
+            ["TotalMonthly"] = "Total/mes",
+            ["PaymentDay"] = "Día de pago",
+            ["AddSource"] = "Agregar fuente",
+            ["AddPaymentDate"] = "Agregar fecha de pago",
+            ["Scheduled"] = "Programado",
+            ["Paid"] = "Pagado",
+            ["Pending"] = "Pendiente",
+            ["PaymentSchedule"] = "Calendario de pagos",
+            ["Analytics"] = "Analítica",
+            ["FinCategories"] = "Por categoría",
+            ["MonthlyTrend"] = "Tendencia mensual",
+            ["ExportFinance"] = "Exportar finanzas",
+            ["AnnualTotal"] = "Anual",
+            ["NextPayment"] = "Próximo pago",
+            ["Date"] = "Fecha",
+            // Finance categories
+            ["CatSalary"] = "Salario",
+            ["CatFreelance"] = "Freelance",
+            ["CatGifts"] = "Regalos",
+            ["CatInvestments"] = "Inversiones",
+            ["CatOtherIncome"] = "Otros ingresos",
+            ["CatFood"] = "Comida",
+            ["CatTransport"] = "Transporte",
+            ["CatHousing"] = "Vivienda",
+            ["CatEntertainment"] = "Entretenimiento",
+            ["CatHealth"] = "Salud",
+            ["CatClothing"] = "Ropa",
+            ["CatSubscriptions"] = "Suscripciones",
+            ["CatEducation"] = "Educación",
+            ["CatOtherExpense"] = "Otros gastos",
+            // Financial Goals
+            ["FinGoals"] = "Metas financieras",
+            ["AddFinGoal"] = "Agregar meta",
+            ["TargetAmount"] = "Monto objetivo",
+            ["SavedAmount"] = "Ahorrado",
+            ["TargetDate"] = "Fecha objetivo",
+            ["GoalReached"] = "¡Meta alcanzada!",
+            ["NoFinGoals"] = "Sin metas financieras",
+            ["GoalProgress"] = "{0}% de la meta",
+            // Accounts
+            ["Accounts"] = "Cuentas",
+            ["AddAccount"] = "Agregar cuenta",
+            ["Transfer"] = "Transferencia",
+            ["AddTransfer"] = "Agregar transferencia",
+            ["TransferFrom"] = "Desde cuenta",
+            ["TransferTo"] = "A cuenta",
+            ["InitialBalance"] = "Saldo inicial",
+            ["AccountBalance"] = "Saldo de cuenta",
+            ["NoAccounts"] = "Sin cuentas",
+            // Filter / Search
+            ["FilterEntries"] = "Filtrar registros",
+            ["SearchPlaceholder"] = "Buscar...",
+            ["AllCategories"] = "Todas las categorías",
+            ["DateFrom"] = "Desde fecha",
+            ["DateTo"] = "Hasta fecha",
+            ["ClearFilter"] = "Limpiar",
+            // Forecast
+            ["Forecast"] = "Pronóstico",
+            ["ForecastBalance"] = "Pronóstico de saldo",
+            ["ProjectedIncome"] = "Ingresos proyectados",
+            ["ProjectedExpenses"] = "Gastos proyectados",
+            ["ForecastDays"] = "por 30 días",
+            // Cashflow
+            ["Cashflow"] = "Flujo de caja",
+            ["CashflowCalendar"] = "Calendario de pagos",
+            // Import
+            ["ImportFile"] = "Importar desde archivo",
+            ["SelectFile"] = "Seleccionar archivo",
+            ["ImportSuccess"] = "{0} registros importados",
+            ["ImportError"] = "Error de importación",
+            ["ImportCSV"] = "Archivos CSV|*.csv",
+            ["ImportExcel"] = "Archivos Excel|*.xlsx",
+            // Split
+            ["SplitTransaction"] = "Dividir operación",
+            ["AddSplit"] = "Agregar parte",
+            ["SplitTotal"] = "Total de partes",
+            ["SplitParent"] = "Registro principal",
+            // Trends
+            ["TrendChart"] = "Tendencias",
+            ["IncomeVsExpenses"] = "Ingresos vs Gastos",
+            ["DueDate"] = "Fecha límite",
+            ["SetDueDate"] = "Establecer fecha límite",
+            ["ClearDate"] = "Borrar fecha",
+            ["PickDate"] = "Elegir fecha",
+
+            // Inbox / Trello
+            ["InboxTitle"] = "Entrada",
+            ["InboxEmpty"] = "No hay tareas en entrada",
+            ["InboxEmptyHint"] = "Añade manualmente o sincroniza con Trello",
+            ["InboxMovedToDate"] = "Tarea movida al {0}",
+            ["AddTask"] = "Añadir",
+            ["SyncTrello"] = "Sincronizar Trello",
+            ["MoveToDay"] = "Mover al día",
+            ["Delete"] = "Eliminar",
+            ["NewTask"] = "Nueva tarea",
+            ["TrelloIntegration"] = "Integración con Trello",
+            ["TrelloDesc"] = "Importa tarjetas de la lista indicada a la bandeja de entrada.",
+            ["TrelloApiKey"] = "Clave API",
+            ["TrelloToken"] = "Token",
+            ["TrelloListName"] = "Nombre de lista (p.ej. \"En proceso\")",
+            ["TrelloHelp"] = "Obtén la clave en https://trello.com/app-key — en la misma página hay un botón \"Token\".",
+            ["TestConnection"] = "Probar conexión",
+            ["TrelloTesting"] = "Probando...",
+            ["TrelloConnectionOk"] = "Conexión correcta",
+            ["TrelloConnectionFail"] = "Falló la conexión",
+            ["TrelloSaved"] = "Configuración guardada",
+            ["TrelloSyncing"] = "Sincronizando...",
+            ["TrelloSyncDone"] = "Añadidas {0} tarjetas",
+            ["TrelloSyncError"] = "Error de sincronización",
+            ["Enabled"] = "Activado",
+            ["Save"] = "Guardar",
+    };
+}
