+++ DailyPlanner/Services/Loc.Fr.cs
@@ -1,0 +1,432 @@
+namespace DailyPlanner.Services;
+
+internal static class LocFrStrings
+{
+    public static Dictionary<string, string> Build() => new()
+    {
+            ["Today"] = "AUJOURD'HUI",
+            ["TodayBtn"] = "Aujourd'hui",
+            ["Statistics"] = "Statistiques",
+            ["Export"] = "Exporter",
+            ["Settings"] = "Paramètres",
+            ["Pomodoro"] = "Pomodoro",
+
+            ["TrayOpen"] = "Ouvrir",
+            ["TrayToday"] = "Aujourd'hui",
+            ["TrayExit"] = "Quitter",
+            ["TrayMinimized"] = "Application réduite dans la barre des tâches",
+
+            ["SearchTooltip"] = "Rechercher (Ctrl+F)",
+            ["CopyWeekTooltip"] = "Copier la structure de la semaine dernière",
+
+            ["NoTasksToday"] = "Pas de tâches aujourd'hui",
+            ["NoTasks"] = "Pas de tâches",
+            ["Completed"] = "terminées",
+
+            ["TotalTasks"] = "Total des tâches",
+            ["CompletedLabel"] = "Terminées",
+            ["NotCompleted"] = "Non terminées",
+            ["GoalsReached"] = "Objectifs atteints",
+            ["TotalTasksTip"] = "Nombre de tâches avec texte pour la semaine",
+            ["CompletedTip"] = "Tâches marquées comme terminées",
+            ["NotCompletedTip"] = "Tâches pas encore terminées",
+            ["GoalsTip"] = "Objectifs atteints sur 4",
+
+            ["WeekGoals"] = "OBJECTIFS DE LA SEMAINE",
+            ["Notes"] = "NOTES",
+            ["Progress"] = "PROGRÈS",
+            ["ProgressLabel"] = "progrès",
+            ["BestDay"] = "meilleur jour",
+            ["GoalsLabel"] = "objectifs",
+            ["WeekTasks"] = "TÂCHES DE LA SEMAINE",
+            ["TodayBadge"] = "AUJOURD'HUI",
+            ["Remaining"] = "restantes",
+            ["CarryOverTip"] = "Reporter les tâches non terminées à demain",
+            ["AddSubTaskTip"] = "Ajouter une sous-tâche",
+            ["DeleteTaskTip"] = "Supprimer la tâche",
+            ["MoveNextDayTip"] = "Déplacer au jour suivant",
+            ["HabitTracker"] = "SUIVI DES HABITUDES",
+            ["StateTracker"] = "BIEN-ÊTRE",
+            ["StateNoData"] = "Suivez votre état chaque jour",
+            ["StateBestDay"] = "Meilleur jour : {0}",
+            ["StateWorstDay"] = "Jour difficile : {0}",
+            ["StateAvg"] = "Moyenne",
+            ["AddNoteTip"] = "Ajouter une note",
+            ["AddGoalTip"] = "Ajouter un objectif",
+            ["DeleteTip"] = "Supprimer",
+            ["AddHabitTip"] = "Ajouter une habitude",
+            ["HabitEmpty"] = "Appuyez sur + pour ajouter une habitude",
+            ["HabitLabel"] = "Habitude",
+            ["ProgressCol"] = "Progrès",
+
+            ["Monday"] = "Lundi", ["Tuesday"] = "Mardi", ["Wednesday"] = "Mercredi",
+            ["Thursday"] = "Jeudi", ["Friday"] = "Vendredi", ["Saturday"] = "Samedi", ["Sunday"] = "Dimanche",
+            ["Mon"] = "Lu", ["Tue"] = "Ma", ["Wed"] = "Me", ["Thu"] = "Je",
+            ["Fri"] = "Ve", ["Sat"] = "Sa", ["Sun"] = "Di",
+
+            ["Sleep"] = "Sommeil", ["Energy"] = "Énergie", ["Mood"] = "Humeur",
+
+            ["Month1"] = "Janvier", ["Month2"] = "Février", ["Month3"] = "Mars",
+            ["Month4"] = "Avril", ["Month5"] = "Mai", ["Month6"] = "Juin",
+            ["Month7"] = "Juillet", ["Month8"] = "Août", ["Month9"] = "Septembre",
+            ["Month10"] = "Octobre", ["Month11"] = "Novembre", ["Month12"] = "Décembre",
+
+            ["PomWork"] = "Travail", ["PomBreak"] = "Pause", ["PomFocus"] = "Focus",
+            ["PomPomodoro"] = "Pomodoro", ["PomFocusMode"] = "Mode focus",
+            ["PomReset"] = "Réinitialiser", ["PomSkip"] = "Passer",
+            ["PomToggleFocus"] = "Basculer Pomodoro/Focus",
+            ["PomSessions"] = "Sessions : ",
+            ["PomPause"] = "\u23F8 Pause", ["PomStart"] = "\u25B6 Démarrer",
+            ["PomBreakTime"] = "Temps de pause ! Repose-toi \U0001f3d6\ufe0f",
+            ["PomWorkTime"] = "Au travail ! \U0001f4aa",
+            ["PomFocusTimer"] = "Minuteur focus",
+            ["PomFocusAlert"] = "Déjà {0} min ! Fais une pause \u2615",
+
+            ["PriorityHigh"] = "Priorité haute (clic pour changer)",
+            ["PriorityMedium"] = "Priorité moyenne (clic pour changer)",
+            ["PriorityLow"] = "Priorité basse (clic pour changer)",
+            ["PriorityNone"] = "Sans priorité",
+
+            ["CatWork"] = "Travail (clic pour changer)", ["CatStudy"] = "Études (clic pour changer)",
+            ["CatPersonal"] = "Personnel (clic pour changer)", ["CatHealthTip"] = "Santé (clic pour changer)",
+            ["CatOther"] = "Autre (clic pour changer)",
+
+            ["SettingsTitle"] = "Paramètres",
+            ["ExportExcel"] = "Exporter vers Excel",
+            ["ExportExcelDesc"] = "Exporter la semaine en cours en fichier .xlsx",
+            ["Palette"] = "Palette",
+            ["PaletteCurrent"] = "Actuelle : ",
+            ["AutoStart"] = "Démarrage automatique",
+            ["AutoStartDesc"] = "Lancer l'application à l'ouverture de Windows",
+            ["Updates"] = "Mises à jour",
+            ["CheckUpdates"] = "Vérifier",
+            ["UpdateBtn"] = "Mettre à jour",
+            ["Backup"] = "Sauvegarde",
+            ["BackupDesc"] = "Créer une copie ou restaurer la base de données",
+            ["CreateBackup"] = "Créer une copie",
+            ["Restore"] = "Restaurer",
+            ["Reminders"] = "Rappels",
+            ["RemindersDesc"] = "Notifications à l'heure indiquée",
+            ["Add"] = "Ajouter",
+            ["TimeTip"] = "Heure (HH:MM)",
+            ["TitleTip"] = "Titre",
+            ["MessageTip"] = "Message",
+            ["RecurringTasks"] = "Tâches récurrentes",
+            ["RecurringDesc"] = "Les modèles sont ajoutés automatiquement chaque semaine",
+            ["AboutApp"] = "À propos",
+            ["Version"] = "Version",
+            ["Platform"] = "Plateforme",
+            ["Database"] = "Base de données",
+            ["DbLocal"] = "SQLite (locale)",
+            ["Storage"] = "Stockage",
+            ["Shortcuts"] = "Raccourcis clavier",
+            ["SwitchMonth"] = "Changer de mois",
+            ["GoToToday"] = "Aller à aujourd'hui",
+            ["SearchTasks"] = "Rechercher des tâches",
+            ["PomodoroTimer"] = "Minuteur Pomodoro",
+            ["ExportExcelShort"] = "Exporter vers Excel",
+            ["OpenSettings"] = "Ouvrir les paramètres",
+            ["Description"] = "Description",
+            ["AppDescription"] = "Daily & Financial Planner — planificateur hebdomadaire avec module financier, suivi des habitudes, analyse de productivité et suivi de l'état. Toutes les données sont stockées localement.",
+            ["Language"] = "Langue",
+            ["LanguageDesc"] = "Langue de l'interface de l'application",
+
+            ["Productivity"] = "Productivité",
+            ["BestDayLabel"] = "Meilleur jour",
+            ["Goals"] = "Objectifs",
+            ["AvgSleep"] = "Moy. sommeil",
+            ["AvgEnergy"] = "Moy. énergie",
+            ["AvgMood"] = "Moy. humeur",
+            ["WeekComparison"] = "Comparaison des semaines",
+            ["PrevWeek"] = "Semaine précédente",
+            ["Trend"] = "Tendance",
+            ["CurrWeek"] = "Semaine en cours",
+            ["WeeklyProductivity"] = "Productivité hebdomadaire",
+            ["HabitHeatmap"] = "Habitudes — Carte thermique (30 jours)",
+            ["HabitStreaks"] = "Séries d'habitudes",
+            ["TasksUnit"] = "tâches",
+            ["NoChange"] = "Pas de changement",
+            ["DaysUnit"] = "j.",
+
+            ["SummaryTasks"] = "Tâches : {0}/{1}",
+            ["SummaryGoals"] = "Objectifs : {0}/{1}",
+            ["SummaryBestDay"] = "Meilleur jour : {0}",
+            ["SummaryExcellent"] = "\u2b50 Excellente semaine !",
+            ["SummaryGood"] = "\u2705 Bonne semaine !",
+            ["SummaryCanBetter"] = "\U0001f4aa On peut faire mieux !",
+
+            ["ExportTitle"] = "Exporter",
+            ["ExportSuccess"] = "Fichier enregistré avec succès",
+            ["ExportError"] = "Erreur lors de l'export",
+            ["RestoreTitle"] = "Restauration",
+            ["RestoreSuccess"] = "Base de données restaurée avec succès",
+            ["RestoreError"] = "Erreur lors de la restauration",
+            ["RestoreInvalidDb"] = "Le fichier n'est pas une base de données SQLite valide",
+            ["CopyTitle"] = "Copier",
+            ["CopySuccess"] = "Structure de la semaine dernière copiée",
+            ["Reminder"] = "Rappel",
+            ["MoveTitle"] = "Déplacer",
+            ["MoveNoSlots"] = "Pas de créneaux libres pour ce jour",
+
+            ["UpdateUnavailable"] = "Mises à jour non disponibles (mode dev)",
+            ["UpdateChecking"] = "Vérification des mises à jour...",
+            ["UpdateLatest"] = "Vous utilisez la dernière version",
+            ["UpdateAvailable"] = "Mise à jour disponible : v{0}",
+            ["UpdateDownloading"] = "Téléchargement de la mise à jour...",
+            ["UpdateProgress"] = "Téléchargement : {0}%",
+
+            ["ExpWeek"] = "Semaine {0}",
+            ["ExpPlannerWeek"] = "Planificateur — Semaine {0} – {1}",
+            ["ExpWeekGoals"] = "Objectifs de la semaine",
+            ["ExpDay"] = "Jour",
+            ["ExpTask"] = "Tâche {0}",
+            ["ExpState"] = "État",
+            ["ExpHabits"] = "Habitudes",
+            ["ExpNotes"] = "Notes",
+
+            ["Loading"] = "Chargement des données...",
+            ["DbError"] = "Erreur d'initialisation de la base de données :\n{0}",
+            ["PlannerFileName"] = "Planificateur_{0}",
+            ["ExcelFilter"] = "Fichiers Excel (*.xlsx)|*.xlsx",
+            ["Goal"] = "Objectif",
+            ["LightTheme"] = "Clair",
+
+            ["Quote1"] = "De petits pas chaque jour mènent à de grands résultats.",
+            ["Quote2"] = "La discipline est le pont entre les objectifs et les réalisations.",
+            ["Quote3"] = "N'attends pas le moment parfait. Prends le moment et rends-le parfait.",
+            ["Quote4"] = "Le succès est la somme de petits efforts répétés jour après jour.",
+            ["Quote5"] = "Planifier, c'est amener le futur dans le présent.",
+            ["Quote6"] = "Chaque jour est une nouvelle opportunité de s'améliorer.",
+            ["Quote7"] = "Fais aujourd'hui ce que les autres ne veulent pas. Demain tu vivras comme ils ne peuvent pas.",
+            ["Quote8"] = "Le focus, ce n'est pas dire 'oui', c'est dire 'non'.",
+            ["Quote9"] = "Les habitudes définissent l'avenir. Choisis-les avec sagesse.",
+            ["Quote10"] = "Le progrès, pas la perfection.",
+            ["Quote11"] = "Le meilleur moment pour commencer, c'est maintenant.",
+            ["Quote12"] = "Ton seul concurrent est celui que tu étais hier.",
+            ["Quote13"] = "L'énergie suit l'attention. Concentre-toi sur l'essentiel.",
+            ["Quote14"] = "Sans plan, tu planifies d'échouer.",
+            ["Quote15"] = "Les grandes choses commencent petit.",
+            ["Quote16"] = "Concentre-toi sur le processus et les résultats viendront.",
+            ["Quote17"] = "Chaque tâche accomplie est une petite victoire.",
+            ["Quote18"] = "Ne remets pas à demain ce qui te rendra meilleur aujourd'hui.",
+            ["Quote19"] = "La productivité n'est pas l'occupation, c'est le résultat.",
+            ["Quote20"] = "Une heure de planification économise dix heures de travail.",
+
+            // Meetings
+            ["Meetings"] = "Réunions",
+            ["MeetingsDesc"] = "Planificateur de réunions avec rappels",
+            ["NewMeeting"] = "Nouvelle réunion",
+            ["MeetingTitle"] = "Titre",
+            ["MeetingDesc"] = "Description",
+            ["MeetingAttendees"] = "Participants",
+            ["MeetingDate"] = "Date",
+            ["MeetingTime"] = "Heure",
+            ["MeetingDuration"] = "Durée",
+            ["MeetingMin"] = "min",
+            ["MeetingNotifyDay"] = "1 jour avant",
+            ["MeetingNotify2h"] = "2 heures avant",
+            ["MeetingNotify30m"] = "30 min avant",
+            ["MeetingTomorrow"] = "Réunion demain",
+            ["MeetingSoon"] = "Réunion bientôt",
+            ["MeetingIn2Hours"] = "dans 2 heures",
+            ["MeetingIn30Min"] = "dans 30 minutes",
+            ["MeetingNoMeetings"] = "Aucune réunion prévue",
+            ["MeetingUpcoming"] = "À venir",
+            ["MeetingPast"] = "Passées",
+
+            ["Undo"] = "Annuler",
+            ["TaskDeleted"] = "Tâche \"{0}\" supprimée",
+            ["CtxDuplicate"] = "Dupliquer",
+            ["CtxPriority"] = "Priorité",
+            ["CtxCategory"] = "Catégorie",
+            ["CtxDeadline"] = "Échéance",
+            ["CtxDeadlineToday"] = "Aujourd'hui",
+            ["CtxDeadlineTomorrow"] = "Demain",
+            ["CtxDeadlineNextWeek"] = "Semaine prochaine",
+            ["CtxDeadlineClear"] = "Supprimer l'échéance",
+            ["CtxMoveNext"] = "Déplacer au jour suivant",
+            ["CtxDelete"] = "Supprimer",
+            ["CtxAddSub"] = "Ajouter sous-tâche",
+            ["CtxNone"] = "Aucun",
+            ["Overdue"] = "En retard",
+            ["DaysShort"] = "j",
+            ["DueToday"] = "Aujourd'hui",
+            ["DueTomorrow"] = "Demain",
+            ["MyDayTitle"] = "Ma journée",
+            ["MyDayGreeting"] = "Bonjour !",
+            ["GreetingMorning"] = "Bonjour !",
+            ["GreetingAfternoon"] = "Bon après-midi !",
+            ["GreetingEvening"] = "Bonsoir !",
+            ["GreetingNight"] = "Bonne nuit !",
+            ["MyDayYesterday"] = "Non terminé hier :",
+            ["MyDayToday"] = "Tâches du jour :",
+            ["MyDayOverdue"] = "Tâches en retard :",
+            ["MyDayNoTasks"] = "Pas de tâches aujourd'hui. C'est l'heure de planifier !",
+            ["MyDayStart"] = "Commencer la journée",
+            ["MyDayDontShow"] = "Ne plus afficher aujourd'hui",
+            ["MyDayEmpty"] = "Nouveau jour — nouvelles opportunités !",
+            ["MyDaySummaryTasks"] = "Tâches : {0} (faites {1})",
+            ["MyDaySummaryIncomplete"] = "Non terminées hier : {0}",
+            ["MyDaySummaryOverdue"] = "En retard : {0}",
+
+            // Finance
+            ["Finance"] = "Finances",
+            ["FinanceTooltip"] = "Finances (Ctrl+M)",
+            ["Income"] = "Revenus",
+            ["Expenses"] = "Dépenses",
+            ["BalanceLabel"] = "Solde",
+            ["Debts"] = "Dettes",
+            ["RecurringPayments"] = "Obligatoires",
+            ["AddEntry"] = "Ajouter une entrée",
+            ["AddDebt"] = "Ajouter une dette",
+            ["AddPayment"] = "Ajouter un paiement",
+            ["AddBudget"] = "Ajouter un budget",
+            ["AddRecurring"] = "Ajouter un paiement",
+            ["Amount"] = "Montant",
+            ["Person"] = "Personne",
+            ["DebtLent"] = "On me doit",
+            ["DebtBorrowed"] = "Je dois",
+            ["DebtSettled"] = "Réglé",
+            ["FinRemaining"] = "Restant",
+            ["Budget"] = "Budget",
+            ["Spent"] = "Dépensé",
+            ["OverBudget"] = "Dépassé",
+            ["FreqMonthly"] = "Mensuel",
+            ["FreqWeekly"] = "Hebdomadaire",
+            ["FreqBiweekly"] = "Bimensuel",
+            ["FreqQuarterly"] = "Trimestriel",
+            ["FreqYearly"] = "Annuel",
+            ["DayOfMonthSuffix"] = "e",
+            ["NoEntries"] = "Aucune entrée",
+            ["NoDebts"] = "Aucune dette",
+            ["FinOverdue"] = "En retard",
+            ["Active"] = "Actif",
+            ["Auto"] = "Auto",
+            ["Remind"] = "Rappel",
+            ["MonthlyObligatory"] = "Obligatoires/mois",
+            ["FreeAmount"] = "Libre",
+            ["Paid"] = "Payé",
+            ["Pending"] = "En attente",
+            ["FinanceDescription"] = "Description",
+            ["Savings"] = "Épargne",
+            ["SavingsRate"] = "Taux d'épargne",
+            ["NetWorth"] = "Valeur nette",
+            ["ChangeIcon"] = "Changer l'icône",
+            ["IncomeSources"] = "Sources de revenus",
+            ["ExpectedIncome"] = "Revenu attendu",
+            ["ReceivedIncome"] = "Reçu",
+            ["ProjectName"] = "Nom du projet",
+            ["ClientName"] = "Client",
+            ["TotalMonthly"] = "Total/mois",
+            ["PaymentDay"] = "Jour de paiement",
+            ["AddSource"] = "Ajouter une source",
+            ["AddPaymentDate"] = "Ajouter une date de paiement",
+            ["Scheduled"] = "Planifié",
+            ["Paid"] = "Payé",
+            ["Pending"] = "En attente",
+            ["PaymentSchedule"] = "Calendrier des paiements",
+            ["Analytics"] = "Analytique",
+            ["FinCategories"] = "Par catégorie",
+            ["MonthlyTrend"] = "Tendance mensuelle",
+            ["ExportFinance"] = "Exporter les finances",
+            ["AnnualTotal"] = "Par an",
+            ["NextPayment"] = "Prochain paiement",
+            ["Date"] = "Date",
+            // Finance categories
+            ["CatSalary"] = "Salaire",
+            ["CatFreelance"] = "Freelance",
+            ["CatGifts"] = "Cadeaux",
+            ["CatInvestments"] = "Investissements",
+            ["CatOtherIncome"] = "Autre revenu",
+            ["CatFood"] = "Alimentation",
+            ["CatTransport"] = "Transport",
+            ["CatHousing"] = "Logement",
+            ["CatEntertainment"] = "Loisirs",
+            ["CatHealth"] = "Santé",
+            ["CatClothing"] = "Vêtements",
+            ["CatSubscriptions"] = "Abonnements",
+            ["CatEducation"] = "Éducation",
+            ["CatOtherExpense"] = "Autre dépense",
+            // Financial Goals
+            ["FinGoals"] = "Objectifs financiers",
+            ["AddFinGoal"] = "Ajouter un objectif",
+            ["TargetAmount"] = "Montant cible",
+            ["SavedAmount"] = "Épargné",
+            ["TargetDate"] = "Date cible",
+            ["GoalReached"] = "Objectif atteint !",
+            ["NoFinGoals"] = "Aucun objectif financier",
+            ["GoalProgress"] = "{0}% de l'objectif",
+            // Accounts
+            ["Accounts"] = "Comptes",
+            ["AddAccount"] = "Ajouter un compte",
+            ["Transfer"] = "Virement",
+            ["AddTransfer"] = "Ajouter un virement",
+            ["TransferFrom"] = "Du compte",
+            ["TransferTo"] = "Vers le compte",
+            ["InitialBalance"] = "Solde initial",
+            ["AccountBalance"] = "Solde du compte",
+            ["NoAccounts"] = "Aucun compte",
+            // Filter / Search
+            ["FilterEntries"] = "Filtrer les entrées",
+            ["SearchPlaceholder"] = "Rechercher...",
+            ["AllCategories"] = "Toutes les catégories",
+            ["DateFrom"] = "À partir du",
+            ["DateTo"] = "Jusqu'au",
+            ["ClearFilter"] = "Effacer",
+            // Forecast
+            ["Forecast"] = "Prévision",
+            ["ForecastBalance"] = "Prévision de solde",
+            ["ProjectedIncome"] = "Revenus prévus",
+            ["ProjectedExpenses"] = "Dépenses prévues",
+            ["ForecastDays"] = "sur 30 jours",
+            // Cashflow
+            ["Cashflow"] = "Flux de trésorerie",
+            ["CashflowCalendar"] = "Calendrier des paiements",
+            // Import
+            ["ImportFile"] = "Importer un fichier",
+            ["SelectFile"] = "Sélectionner un fichier",
+            ["ImportSuccess"] = "{0} entrées importées",
+            ["ImportError"] = "Erreur d'importation",
+            ["ImportCSV"] = "Fichiers CSV|*.csv",
+            ["ImportExcel"] = "Fichiers Excel|*.xlsx",
+            // Split
+            ["SplitTransaction"] = "Diviser l'opération",
+            ["AddSplit"] = "Ajouter une partie",
+            ["SplitTotal"] = "Total des parties",
+            ["SplitParent"] = "Entrée principale",
+            // Trends
+            ["TrendChart"] = "Tendances",
+            ["IncomeVsExpenses"] = "Revenus vs Dépenses",
+            ["DueDate"] = "Échéance",
+            ["SetDueDate"] = "Définir l'échéance",
+            ["ClearDate"] = "Effacer la date",
+            ["PickDate"] = "Choisir une date",
+
+            // Inbox / Trello
+            ["InboxTitle"] = "Boîte de réception",
+            ["InboxEmpty"] = "Aucune tâche entrante",
+            ["InboxEmptyHint"] = "Ajoutez manuellement ou synchronisez Trello",
+            ["InboxMovedToDate"] = "Tâche déplacée au {0}",
+            ["AddTask"] = "Ajouter",
+            ["SyncTrello"] = "Synchroniser Trello",
+            ["MoveToDay"] = "Déplacer au jour",
+            ["Delete"] = "Supprimer",
+            ["NewTask"] = "Nouvelle tâche",
+            ["TrelloIntegration"] = "Intégration Trello",
+            ["TrelloDesc"] = "Importe les cartes de la liste indiquée dans la boîte de réception.",
+            ["TrelloApiKey"] = "Clé API",
+            ["TrelloToken"] = "Jeton",
+            ["TrelloListName"] = "Nom de la liste (ex. «En cours»)",
+            ["TrelloHelp"] = "Obtenez la clé sur https://trello.com/app-key — le bouton «Token» est sur la même page.",
+            ["TestConnection"] = "Tester la connexion",
+            ["TrelloTesting"] = "Test...",
+            ["TrelloConnectionOk"] = "Connexion réussie",
+            ["TrelloConnectionFail"] = "Échec de la connexion",
+            ["TrelloSaved"] = "Paramètres enregistrés",
+            ["TrelloSyncing"] = "Synchronisation...",
+            ["TrelloSyncDone"] = "{0} cartes ajoutées",
+            ["TrelloSyncError"] = "Erreur de synchronisation",
+            ["Enabled"] = "Activé",
+            ["Save"] = "Enregistrer",
+    };
+}
