+++ DailyPlanner/Services/Loc.Ru.cs
@@ -1,0 +1,484 @@
+namespace DailyPlanner.Services;
+
+internal static class LocRuStrings
+{
+    public static Dictionary<string, string> Build() => new()
+    {
+            // Sidebar
+            ["Today"] = "СЕГОДНЯ",
+            ["TodayBtn"] = "Сегодня",
+            ["Statistics"] = "Статистика",
+            ["Export"] = "Экспорт",
+            ["Settings"] = "Настройки",
+            ["Pomodoro"] = "Pomodoro",
+
+            // Tray
+            ["TrayOpen"] = "Открыть",
+            ["TrayToday"] = "Сегодня",
+            ["TrayExit"] = "Выход",
+            ["TrayMinimized"] = "Приложение свёрнуто в трей",
+
+            // Search
+            ["SearchTooltip"] = "Поиск (Ctrl+F)",
+            ["CopyWeekTooltip"] = "Копировать структуру прошлой недели",
+
+            // Today widget
+            ["NoTasksToday"] = "Нет задач на сегодня",
+            ["NoTasks"] = "Нет задач",
+            ["Completed"] = "выполнено",
+
+            // Stat cards (week)
+            ["TotalTasks"] = "Всего задач",
+            ["CompletedLabel"] = "Выполнено",
+            ["NotCompleted"] = "Не выполнено",
+            ["GoalsReached"] = "Целей достигнуто",
+            ["TotalTasksTip"] = "Количество задач с текстом за неделю",
+            ["CompletedTip"] = "Задач, отмеченных как выполненные",
+            ["NotCompletedTip"] = "Задач, которые ещё не выполнены",
+            ["GoalsTip"] = "Выполненных целей из 4",
+
+            // Sections
+            ["WeekGoals"] = "ЦЕЛИ НЕДЕЛИ",
+            ["Notes"] = "ЗАМЕТКИ",
+            ["Progress"] = "ПРОГРЕСС",
+            ["ProgressLabel"] = "прогресс",
+            ["BestDay"] = "лучший день",
+            ["GoalsLabel"] = "целей",
+            ["WeekTasks"] = "ЗАДАЧИ НА НЕДЕЛЮ",
+            ["TodayBadge"] = "СЕГОДНЯ",
+            ["Remaining"] = "осталось",
+            ["CarryOverTip"] = "Перенести незавершённые на завтра",
+            ["AddSubTaskTip"] = "Добавить подзадачу",
+            ["DeleteTaskTip"] = "Удалить задачу",
+            ["MoveNextDayTip"] = "Перенести на следующий день",
+            ["HabitTracker"] = "ТРЕКЕР ПРИВЫЧЕК",
+            ["StateTracker"] = "САМОЧУВСТВИЕ",
+            ["StateNoData"] = "Отмечайте состояние каждый день",
+            ["StateBestDay"] = "Лучший день: {0}",
+            ["StateWorstDay"] = "Тяжёлый день: {0}",
+            ["StateAvg"] = "Среднее",
+            ["AddNoteTip"] = "Добавить заметку",
+            ["AddGoalTip"] = "Добавить цель",
+            ["DeleteTip"] = "Удалить",
+            ["AddHabitTip"] = "Добавить привычку",
+            ["HabitEmpty"] = "Нажмите + чтобы добавить привычку",
+            ["HabitLabel"] = "Привычка",
+            ["ProgressCol"] = "Прогресс",
+
+            // Day names
+            ["Monday"] = "Понедельник",
+            ["Tuesday"] = "Вторник",
+            ["Wednesday"] = "Среда",
+            ["Thursday"] = "Четверг",
+            ["Friday"] = "Пятница",
+            ["Saturday"] = "Суббота",
+            ["Sunday"] = "Воскресенье",
+            ["Mon"] = "Пн",
+            ["Tue"] = "Вт",
+            ["Wed"] = "Ср",
+            ["Thu"] = "Чт",
+            ["Fri"] = "Пт",
+            ["Sat"] = "Сб",
+            ["Sun"] = "Вс",
+
+            // State columns
+            ["Sleep"] = "Сон",
+            ["Energy"] = "Энергия",
+            ["Mood"] = "Настрой",
+
+            // Month names
+            ["Month1"] = "Январь", ["Month2"] = "Февраль", ["Month3"] = "Март",
+            ["Month4"] = "Апрель", ["Month5"] = "Май", ["Month6"] = "Июнь",
+            ["Month7"] = "Июль", ["Month8"] = "Август", ["Month9"] = "Сентябрь",
+            ["Month10"] = "Октябрь", ["Month11"] = "Ноябрь", ["Month12"] = "Декабрь",
+
+            // Pomodoro
+            ["PomWork"] = "Работа",
+            ["PomBreak"] = "Перерыв",
+            ["PomFocus"] = "Фокус",
+            ["PomPomodoro"] = "Помодоро",
+            ["PomFocusMode"] = "Фокус-режим",
+            ["PomReset"] = "Сброс",
+            ["PomSkip"] = "Пропустить",
+            ["PomToggleFocus"] = "Переключить Помодоро/Фокус",
+            ["PomSessions"] = "Сессий: ",
+            ["PomPause"] = "\u23F8 Пауза",
+            ["PomStart"] = "\u25B6 Старт",
+            ["PomBreakTime"] = "Время перерыва! Отдохни \U0001f3d6\ufe0f",
+            ["PomWorkTime"] = "Время работать! \U0001f4aa",
+            ["PomFocusTimer"] = "Фокус-таймер",
+            ["PomFocusAlert"] = "Уже {0} мин! Сделай перерыв \u2615",
+
+            // Priority tooltips
+            ["PriorityHigh"] = "Высокий приоритет (клик — сменить)",
+            ["PriorityMedium"] = "Средний приоритет (клик — сменить)",
+            ["PriorityLow"] = "Низкий приоритет (клик — сменить)",
+            ["PriorityNone"] = "Без приоритета",
+
+            // Category tooltips
+            ["CatWork"] = "Работа (клик — сменить)",
+            ["CatStudy"] = "Учёба (клик — сменить)",
+            ["CatPersonal"] = "Личное (клик — сменить)",
+            ["CatHealthTip"] = "Здоровье (клик — сменить)",
+            ["CatOther"] = "Другое (клик — сменить)",
+
+            // Settings
+            ["SettingsTitle"] = "Настройки",
+            ["ExportExcel"] = "Экспорт в Excel",
+            ["ExportExcelDesc"] = "Выгрузить текущую неделю в .xlsx файл",
+            ["Palette"] = "Палитра",
+            ["PaletteCurrent"] = "Текущая: ",
+            ["AutoStart"] = "Автозапуск",
+            ["AutoStartDesc"] = "Запускать приложение при входе в Windows",
+            ["Updates"] = "Обновления",
+            ["CheckUpdates"] = "Проверить",
+            ["UpdateBtn"] = "Обновить",
+            ["Backup"] = "Резервное копирование",
+            ["BackupDesc"] = "Создайте копию или восстановите базу данных",
+            ["CreateBackup"] = "Создать копию",
+            ["Restore"] = "Восстановить",
+            ["Reminders"] = "Напоминания",
+            ["RemindersDesc"] = "Уведомления в заданное время",
+            ["Add"] = "Добавить",
+            ["TimeTip"] = "Время (ЧЧ:ММ)",
+            ["TitleTip"] = "Заголовок",
+            ["MessageTip"] = "Сообщение",
+            ["RecurringTasks"] = "Повторяющиеся задачи",
+            ["RecurringDesc"] = "Шаблоны автоматически добавляются каждую неделю",
+            ["AboutApp"] = "О приложении",
+            ["Version"] = "Версия",
+            ["Platform"] = "Платформа",
+            ["Database"] = "База данных",
+            ["DbLocal"] = "SQLite (локальная)",
+            ["Storage"] = "Хранение",
+            ["Shortcuts"] = "Горячие клавиши",
+            ["SwitchMonth"] = "Переключить месяц",
+            ["GoToToday"] = "Перейти к сегодня",
+            ["SearchTasks"] = "Поиск по задачам",
+            ["PomodoroTimer"] = "Помодоро таймер",
+            ["ExportExcelShort"] = "Экспорт в Excel",
+            ["OpenSettings"] = "Открыть настройки",
+            ["Description"] = "Описание",
+            ["AppDescription"] = "Daily & Financial Planner — недельный планер с модулем финансов, трекером привычек, аналитикой продуктивности и мониторингом состояния. Все данные хранятся локально.",
+            ["Language"] = "Язык",
+            ["LanguageDesc"] = "Язык интерфейса приложения",
+
+            // Statistics
+            ["Productivity"] = "Продуктивность",
+            ["BestDayLabel"] = "Лучший день",
+            ["Goals"] = "Цели",
+            ["AvgSleep"] = "Ср. сон",
+            ["AvgEnergy"] = "Ср. энергия",
+            ["AvgMood"] = "Ср. настроение",
+            ["WeekComparison"] = "Сравнение недель",
+            ["PrevWeek"] = "Прошлая неделя",
+            ["Trend"] = "Тренд",
+            ["CurrWeek"] = "Текущая неделя",
+            ["WeeklyProductivity"] = "Продуктивность по неделям",
+            ["HabitHeatmap"] = "Привычки — Heatmap (30 дней)",
+            ["HabitStreaks"] = "Серии привычек",
+            ["TasksUnit"] = "задач",
+            ["NoChange"] = "Без изменений",
+            ["DaysUnit"] = "дн.",
+
+            // Weekly summary
+            ["SummaryTasks"] = "Задачи: {0}/{1}",
+            ["SummaryGoals"] = "Цели: {0}/{1}",
+            ["SummaryBestDay"] = "Лучший день: {0}",
+            ["SummaryExcellent"] = "\u2b50 Отличная неделя!",
+            ["SummaryGood"] = "\u2705 Хорошая неделя!",
+            ["SummaryCanBetter"] = "\U0001f4aa Можно лучше!",
+
+            // Notifications/toast
+            ["ExportTitle"] = "Экспорт",
+            ["ExportSuccess"] = "Файл успешно сохранён",
+            ["ExportError"] = "Ошибка при экспорте",
+            ["RestoreTitle"] = "Восстановление",
+            ["RestoreSuccess"] = "База данных успешно восстановлена",
+            ["RestoreError"] = "Ошибка при восстановлении",
+            ["RestoreInvalidDb"] = "Файл не является базой данных SQLite",
+            ["CopyTitle"] = "Копирование",
+            ["CopySuccess"] = "Структура прошлой недели скопирована",
+            ["Reminder"] = "Напоминание",
+            ["MoveTitle"] = "Перемещение",
+            ["MoveNoSlots"] = "Нет свободных слотов в этом дне",
+
+            // Updates
+            ["UpdateUnavailable"] = "Обновления недоступны (dev-режим)",
+            ["UpdateChecking"] = "Проверка обновлений...",
+            ["UpdateLatest"] = "Вы используете последнюю версию",
+            ["UpdateAvailable"] = "Доступно обновление: v{0}",
+            ["UpdateDownloading"] = "Скачивание обновления...",
+            ["UpdateProgress"] = "Скачивание: {0}%",
+
+            // Export service
+            ["ExpWeek"] = "Неделя {0}",
+            ["ExpPlannerWeek"] = "Планер — Неделя {0} – {1}",
+            ["ExpWeekGoals"] = "Цели недели",
+            ["ExpDay"] = "День",
+            ["ExpTask"] = "Задача {0}",
+            ["ExpState"] = "Состояние",
+            ["ExpHabits"] = "Привычки",
+            ["ExpNotes"] = "Заметки",
+
+            // Splash / errors
+            ["Loading"] = "Загрузка данных...",
+            ["DbError"] = "Ошибка инициализации базы данных:\n{0}",
+
+            // Save dialog
+            ["PlannerFileName"] = "Планер_{0}",
+            ["ExcelFilter"] = "Excel файлы (*.xlsx)|*.xlsx",
+
+            // Search context
+            ["Goal"] = "Цель",
+
+            // Light theme
+            ["LightTheme"] = "Светлая",
+
+            // Quotes
+            ["Quote1"] = "Маленькие шаги каждый день приводят к большим результатам.",
+            ["Quote2"] = "Дисциплина — это мост между целями и достижениями.",
+            ["Quote3"] = "Не жди идеального момента. Бери момент и делай его идеальным.",
+            ["Quote4"] = "Успех — это сумма маленьких усилий, повторяемых изо дня в день.",
+            ["Quote5"] = "Планирование — это приведение будущего в настоящее.",
+            ["Quote6"] = "Каждый день — это новая возможность стать лучше.",
+            ["Quote7"] = "Делай сегодня то, что другие не хотят. Завтра будешь жить так, как другие не могут.",
+            ["Quote8"] = "Фокус — это не то, чему ты говоришь 'да', а то, чему говоришь 'нет'.",
+            ["Quote9"] = "Привычки определяют будущее. Выбирай их мудро.",
+            ["Quote10"] = "Прогресс, а не совершенство.",
+            ["Quote11"] = "Лучшее время начать — прямо сейчас.",
+            ["Quote12"] = "Твой единственный конкурент — ты вчерашний.",
+            ["Quote13"] = "Энергия следует за вниманием. Направь внимание на важное.",
+            ["Quote14"] = "Без плана ты планируешь потерпеть неудачу.",
+            ["Quote15"] = "Великие дела начинаются с малого.",
+            ["Quote16"] = "Сосредоточься на процессе, а результат придёт.",
+            ["Quote17"] = "Каждая выполненная задача — маленькая победа.",
+            ["Quote18"] = "Не откладывай на завтра то, что сделает тебя лучше сегодня.",
+            ["Quote19"] = "Продуктивность — это не занятость, а результат.",
+            ["Quote20"] = "Один час планирования экономит десять часов работы.",
+
+            // Meetings
+            ["Meetings"] = "Встречи",
+            ["MeetingsDesc"] = "Планировщик встреч с напоминаниями",
+            ["NewMeeting"] = "Новая встреча",
+            ["MeetingTitle"] = "Название",
+            ["MeetingDesc"] = "Описание",
+            ["MeetingAttendees"] = "Участники",
+            ["MeetingDate"] = "Дата",
+            ["MeetingTime"] = "Время",
+            ["MeetingDuration"] = "Длительность",
+            ["MeetingMin"] = "мин",
+            ["MeetingNotifyDay"] = "За день",
+            ["MeetingNotify2h"] = "За 2 часа",
+            ["MeetingNotify30m"] = "За 30 мин",
+            ["MeetingTomorrow"] = "Встреча завтра",
+            ["MeetingSoon"] = "Встреча скоро",
+            ["MeetingIn2Hours"] = "через 2 часа",
+            ["MeetingIn30Min"] = "через 30 минут",
+            ["MeetingNoMeetings"] = "Нет запланированных встреч",
+            ["MeetingUpcoming"] = "Предстоящие",
+            ["MeetingPast"] = "Прошедшие",
+
+            // Undo
+            ["Undo"] = "Отменить",
+            ["TaskDeleted"] = "Задача \"{0}\" удалена",
+
+            // Context menu
+            ["CtxDuplicate"] = "Дублировать",
+            ["CtxPriority"] = "Приоритет",
+            ["CtxCategory"] = "Категория",
+            ["CtxDeadline"] = "Дедлайн",
+            ["CtxDeadlineToday"] = "Сегодня",
+            ["CtxDeadlineTomorrow"] = "Завтра",
+            ["CtxDeadlineNextWeek"] = "Через неделю",
+            ["CtxDeadlineClear"] = "Убрать дедлайн",
+            ["CtxMoveNext"] = "На следующий день",
+            ["CtxDelete"] = "Удалить",
+            ["CtxAddSub"] = "Добавить подзадачу",
+            ["CtxNone"] = "Нет",
+
+            // Deadline
+            ["Overdue"] = "Просрочено",
+            ["DaysShort"] = "д.",
+            ["DueToday"] = "Сегодня",
+            ["DueTomorrow"] = "Завтра",
+
+            // My Day
+            ["MyDayTitle"] = "Мой день",
+            ["MyDayGreeting"] = "Доброе утро!",
+            ["GreetingMorning"] = "Доброе утро!",
+            ["GreetingAfternoon"] = "Добрый день!",
+            ["GreetingEvening"] = "Добрый вечер!",
+            ["GreetingNight"] = "Доброй ночи!",
+            ["MyDayYesterday"] = "Вчера не завершено:",
+            ["MyDayToday"] = "Задачи на сегодня:",
+            ["MyDayOverdue"] = "Просроченные задачи:",
+            ["MyDayNoTasks"] = "Нет задач на сегодня. Время планировать!",
+            ["MyDayStart"] = "Начать день",
+            ["MyDayDontShow"] = "Не показывать сегодня",
+            ["MyDayEmpty"] = "Новый день — новые возможности!",
+            ["MyDaySummaryTasks"] = "Задач: {0} (выполнено {1})",
+            ["MyDaySummaryIncomplete"] = "Не завершено вчера: {0}",
+            ["MyDaySummaryOverdue"] = "Просрочено: {0}",
+
+            // Finance
+            ["Finance"] = "Финансы",
+            ["FinanceTooltip"] = "Финансы (Ctrl+M)",
+            ["Income"] = "Доходы",
+            ["Expenses"] = "Расходы",
+            ["BalanceLabel"] = "Баланс",
+            ["Debts"] = "Долги",
+            ["RecurringPayments"] = "Обязательные",
+            ["AddEntry"] = "Добавить запись",
+            ["AddDebt"] = "Добавить долг",
+            ["AddPayment"] = "Добавить платёж",
+            ["AddBudget"] = "Добавить бюджет",
+            ["AddRecurring"] = "Добавить платёж",
+            ["Amount"] = "Сумма",
+            ["Person"] = "Кому / от кого",
+            ["DebtLent"] = "Мне должны",
+            ["DebtBorrowed"] = "Я должен",
+            ["DebtSettled"] = "Погашен",
+            ["FinRemaining"] = "Остаток",
+            ["Budget"] = "Бюджет",
+            ["Spent"] = "Потрачено",
+            ["OverBudget"] = "Превышен",
+            ["FreqMonthly"] = "Ежемесячно",
+            ["FreqWeekly"] = "Еженедельно",
+            ["FreqBiweekly"] = "Раз в 2 недели",
+            ["FreqQuarterly"] = "Ежеквартально",
+            ["FreqYearly"] = "Ежегодно",
+            ["DayOfMonthSuffix"] = "е число",
+            ["NoEntries"] = "Нет записей",
+            ["NoDebts"] = "Нет долгов",
+            ["FinOverdue"] = "Просрочен",
+            ["Active"] = "Актив.",
+            ["Auto"] = "Авто",
+            ["Remind"] = "Напом.",
+            ["MonthlyObligatory"] = "Обязательные/мес",
+            ["FreeAmount"] = "Свободно",
+            ["Paid"] = "Оплачено",
+            ["Pending"] = "Ожидает",
+            ["FinanceDescription"] = "Описание",
+            ["Savings"] = "Накопления",
+            ["SavingsRate"] = "Норма сбережений",
+            ["NetWorth"] = "Чистый баланс",
+            ["ChangeIcon"] = "Сменить значок",
+            ["IncomeSources"] = "Источники дохода",
+            ["ExpectedIncome"] = "Ожидаемый доход",
+            ["ReceivedIncome"] = "Получено",
+            ["ProjectName"] = "Название проекта",
+            ["ClientName"] = "Клиент",
+            ["TotalMonthly"] = "Сумма/мес",
+            ["PaymentDay"] = "День оплаты",
+            ["AddSource"] = "Добавить источник",
+            ["AddPaymentDate"] = "Добавить дату оплаты",
+            ["Scheduled"] = "Запланировано",
+            ["Paid"] = "Оплачено",
+            ["Pending"] = "Ожидается",
+            ["PaymentSchedule"] = "График оплат",
+            ["Analytics"] = "Аналитика",
+            ["FinCategories"] = "По категориям",
+            ["MonthlyTrend"] = "Тренд по месяцам",
+            ["ExportFinance"] = "Экспорт финансов",
+            ["AnnualTotal"] = "В год",
+            ["NextPayment"] = "Следующий платёж",
+            ["Date"] = "Дата",
+            // Finance categories
+            ["CatSalary"] = "Зарплата",
+            ["CatFreelance"] = "Фриланс",
+            ["CatGifts"] = "Подарки",
+            ["CatInvestments"] = "Инвестиции",
+            ["CatOtherIncome"] = "Прочий доход",
+            ["CatFood"] = "Еда",
+            ["CatTransport"] = "Транспорт",
+            ["CatHousing"] = "Жильё",
+            ["CatEntertainment"] = "Развлечения",
+            ["CatHealth"] = "Здоровье",
+            ["CatClothing"] = "Одежда",
+            ["CatSubscriptions"] = "Подписки",
+            ["CatEducation"] = "Образование",
+            ["CatOtherExpense"] = "Прочий расход",
+            // Financial Goals
+            ["FinGoals"] = "Финансовые цели",
+            ["AddFinGoal"] = "Добавить цель",
+            ["TargetAmount"] = "Целевая сумма",
+            ["SavedAmount"] = "Накоплено",
+            ["TargetDate"] = "Дата цели",
+            ["GoalReached"] = "Цель достигнута!",
+            ["NoFinGoals"] = "Нет финансовых целей",
+            ["GoalProgress"] = "{0}% от цели",
+            // Accounts
+            ["Accounts"] = "Счета",
+            ["AddAccount"] = "Добавить счёт",
+            ["Transfer"] = "Перевод",
+            ["AddTransfer"] = "Добавить перевод",
+            ["TransferFrom"] = "Со счёта",
+            ["TransferTo"] = "На счёт",
+            ["InitialBalance"] = "Начальный баланс",
+            ["AccountBalance"] = "Баланс счёта",
+            ["NoAccounts"] = "Нет счетов",
+            // Filter / Search
+            ["FilterEntries"] = "Фильтр записей",
+            ["SearchPlaceholder"] = "Поиск...",
+            ["AllCategories"] = "Все категории",
+            ["DateFrom"] = "С даты",
+            ["DateTo"] = "По дату",
+            ["ClearFilter"] = "Сбросить",
+            // Forecast
+            ["Forecast"] = "Прогноз",
+            ["ForecastBalance"] = "Прогноз баланса",
+            ["ProjectedIncome"] = "Ожидаемые доходы",
+            ["ProjectedExpenses"] = "Ожидаемые расходы",
+            ["ForecastDays"] = "на 30 дней",
+            // Cashflow
+            ["Cashflow"] = "Денежный поток",
+            ["CashflowCalendar"] = "Календарь платежей",
+            // Import
+            ["ImportFile"] = "Импорт из файла",
+            ["SelectFile"] = "Выбрать файл",
+            ["ImportSuccess"] = "Импортировано {0} записей",
+            ["ImportError"] = "Ошибка импорта",
+            ["ImportCSV"] = "CSV файлы|*.csv",
+            ["ImportExcel"] = "Excel файлы|*.xlsx",
+            // Split
+            ["SplitTransaction"] = "Разделить операцию",
+            ["AddSplit"] = "Добавить часть",
+            ["SplitTotal"] = "Итого частей",
+            ["SplitParent"] = "Основная запись",
+            // Trends
+            ["TrendChart"] = "Тренды",
+            ["IncomeVsExpenses"] = "Доходы vs Расходы",
+            ["DueDate"] = "Срок",
+            ["SetDueDate"] = "Указать срок",
+            ["ClearDate"] = "Убрать дату",
+            ["PickDate"] = "Выбрать дату",
+
+            // Inbox / Trello
+            ["InboxTitle"] = "Входящие",
+            ["InboxEmpty"] = "Нет входящих задач",
+            ["InboxEmptyHint"] = "Добавьте вручную или синхронизируйте Trello",
+            ["InboxMovedToDate"] = "Задача перемещена на {0}",
+            ["AddTask"] = "Добавить",
+            ["SyncTrello"] = "Синхронизировать Trello",
+            ["MoveToDay"] = "Перенести в день",
+            ["Delete"] = "Удалить",
+            ["NewTask"] = "Новая задача",
+            ["TrelloIntegration"] = "Интеграция с Trello",
+            ["TrelloDesc"] = "Подтягивает карточки из указанного списка во входящие планера.",
+            ["TrelloApiKey"] = "API ключ",
+            ["TrelloToken"] = "Токен",
+            ["TrelloListName"] = "Название списка (например, «В работе»)",
+            ["TrelloHelp"] = "Получить ключ: https://trello.com/app-key — там же создаётся токен по кнопке «Token».",
+            ["TestConnection"] = "Проверить соединение",
+            ["TrelloTesting"] = "Проверка...",
+            ["TrelloConnectionOk"] = "Соединение успешно",
+            ["TrelloConnectionFail"] = "Не удалось подключиться",
+            ["TrelloSaved"] = "Настройки сохранены",
+            ["TrelloSyncing"] = "Синхронизация...",
+            ["TrelloSyncDone"] = "Добавлено карточек: {0}",
+            ["TrelloSyncError"] = "Ошибка синхронизации",
+            ["Enabled"] = "Включено",
+            ["Save"] = "Сохранить",
+    };
+}
